Cross-browser with flex
I set the parent to display:flex. I wrapped the image in a div and set the div and following paragraph (flex children) to a max-wdith of 500px. I then set the width of the image to 100%. This made the image and text look the same in Firefox, Chrome, and Sarafi.